feat: v3.6.0-rc4 (#207)
* feat: oauth reform for potential improvements

* fix: update avatars with new pfp

* fix: remove redundant include

* feat: v3.6.0-rc4


* fix: remove console log

+++ src/pages/api/auth/oauth/index.ts
@@ -1,22 +1,42 @@
 import prisma from 'lib/prisma';
 import { NextApiReq, NextApiRes, withZipline } from 'lib/middleware/withZipline';
+import { OauthProviders } from '@prisma/client';
 
 async function handler(req: NextApiReq, res: NextApiRes) {
   const user = await req.user();
   if (!user) return res.error('not logged in');
 
   if (req.method === 'DELETE') {
-    if (!user.password)
+    if (!user.password && user.oauth.length === 1)
       return res.forbid("can't unlink account without a password, please set one then unlink.");
 
+    const { provider } = req.body as { provider: OauthProviders };
+
+    if (!provider) {
+      const nuser = await prisma.user.update({
+        where: { id: user.id },
+        data: {
+          oauth: {
+            deleteMany: {},
+          },
+        },
+      });
+
+      delete nuser.password;
+
+      return res.json(nuser);
+    }
     const nuser = await prisma.user.update({
       where: {
         id: user.id,
       },
       data: {
-        oauth: false,
-        oauthProvider: null,
-        oauthAccessToken: null,
+        oauth: {
+          deleteMany: [{ provider }],
+        },
+      },
+      include: {
+        oauth: true,
       },
     });
 
@@ -24,11 +44,7 @@ async function handler(req: NextApiReq, res: NextApiRes) {
 
     return res.json(nuser);
   } else {
-    return res.json({
-      enabled: user.oauth,
-      provider: user.oauthProvider,
-      access_token: user.oauthAccessToken,
-    });
+    return res.json(user.oauth);
   }
 }
